Distinct the elements in DE solution

+++ Algorithm/DifferentialEvolutions/Individual.cs
@@ -5,16 +5,16 @@ namespace SLAPScheduling.Algorithm.DifferentialEvolutions
     public class Individual
     {
         public double Fitness { get; set; }
-        public List<double> Elements { get; set; }
+        public List<int> Elements { get; set; }
 
-        public Individual(List<double> Elements)
+        public Individual(List<int> Elements)
         {
             this.Elements = Elements;
         }
 
         public Individual()
         {
-            Elements = new List<double>();
+            Elements = new List<int>();
         }
 
         public double Evaluate(List<ReceiptSublot> receiptSubLots, Dictionary<int, Location> locationDictionary, int numberOfReceiptSubLots)
@@ -29,8 +29,9 @@ namespace SLAPScheduling.Algorithm.DifferentialEvolutions
             StringBuilder s = new StringBuilder();
             foreach (var item in Elements)
             {
-                s.Append(Math.Round(item, 4) + ", ");
+                s.Append(Math.Round((decimal)item, 4) + ", ");
             }
+
             s.Remove(s.Length - 2, 2);
             return s.ToString();
         }
